solve second puzzle correctly

+++ 2024/Day21/Day21.Logic/CombinedKeypadTyping.cs
@@ -20,7 +20,7 @@ public class CombinedKeypadTyping : IKeypadTyping
         return currentSequence;
     }
 
-    public void CountShortestSequence(string sequenceToType, Dictionary<string, int> memoization)
+    public void CountShortestSequence(string sequenceToType, Dictionary<string, long> memoization)
     {
         _keypads.First().CountShortestSequence(sequenceToType, memoization);
         foreach (var keypad in _keypads.Skip(1))
